Finished final article.
This is the last meaningful commit that this repository will see for the forseeable future, so long as the submissions are without errors. The state of the project is regrettable - the ongoing coronavirus pandemic negated the chance of any real-life work and building ocurring. Indeed, it was quite lucky that the speaker was built when it was. My thanks to my project supervisor for helping me through these past 8 months. I would like to finish the project off in my spare time, when I have access to the right equipment to make that happen. This will at the very earliest be around September 2021 after my Year in Industry. For those who come across this project, feel free to continue where I left off.